refactor(core): Wave 4 — split 5 files with design pattern discipline
Apply Martin Fowler patterns, not mechanical line-count splits:

- dual_precision.rs (538→377): Extract Class → int8_traversal.rs (186)
  follows existing rabitq_traversal.rs pattern; fixes fn NLOC 59→41
- match_clause.rs (537→274): Decompose Conditional → match_patterns.rs (289)
  CC reduced from 13→7 via Extract Function (scan_outside_quotes)
- execution_paths.rs (521→405): Strategy → metadata_query.rs (110)
  + DRY: removed duplicate scan_candidate_ids_metadata
- edge.rs (516→416): Extract Class → edge_removal.rs (110) by domain
- crud.rs (516→481): Move QuantizationGuards to crud_helpers (existing SRP home)

All files under 500 NLOC. Zero public API changes. CC <= 8 everywhere.
