+++ crates/velesdb-core/src/collection/core/crud.rs
@@ -2,53 +2,18 @@
 //!
 //! Read and delete operations are in `crud_read_delete.rs`.
 //! Bulk-specific methods (`upsert_bulk`, `upsert_bulk_from_raw`) are in `crud_bulk.rs`.
-//! Quantization caching helpers and secondary-index update helpers are in `crud_helpers.rs`.
+//! Quantization caching helpers, secondary-index update helpers,
+//! `DedupMap`, and `QuantizationGuards` are in `crud_helpers.rs`.
 
+use super::crud_helpers::{DedupMap, QuantizationGuards};
 use crate::collection::types::Collection;
 use crate::error::{Error, Result};
 use crate::point::Point;
-use crate::quantization::{BinaryQuantizedVector, PQVector, QuantizedVector, StorageMode};
+use crate::quantization::StorageMode;
 use crate::storage::{LogPayloadStorage, PayloadStorage, VectorStorage};
 use crate::validation::validate_dimension_match;
 
-use parking_lot::RwLockWriteGuard;
 use std::collections::{BTreeMap, HashMap};
-
-/// Pre-computed last-writer-wins dedup map: `point_id -> index_of_last_occurrence`.
-///
-/// Built once in `batch_store_all` and shared by both `write_deduped_payloads`
-/// and `write_deduped_vectors` to avoid redundant map construction (Issue #425).
-pub(super) type DedupMap = HashMap<u64, usize>;
-
-pub(super) struct QuantizationGuards<'a> {
-    pub(super) sq8: Option<RwLockWriteGuard<'a, HashMap<u64, QuantizedVector>>>,
-    pub(super) binary: Option<RwLockWriteGuard<'a, HashMap<u64, BinaryQuantizedVector>>>,
-    pub(super) pq: Option<RwLockWriteGuard<'a, HashMap<u64, PQVector>>>,
-}
-
-impl<'a> QuantizationGuards<'a> {
-    fn acquire(collection: &'a Collection, mode: StorageMode) -> Self {
-        Self {
-            sq8: matches!(mode, StorageMode::SQ8).then(|| collection.sq8_cache.write()),
-            binary: matches!(mode, StorageMode::Binary).then(|| collection.binary_cache.write()),
-            pq: matches!(mode, StorageMode::ProductQuantization)
-                .then(|| collection.pq_cache.write()),
-        }
-    }
-
-    /// Acquires only the PQ cache guard (for when SQ8/Binary were handled in parallel).
-    ///
-    /// Issue #486: After parallel quantization for SQ8/Binary, only PQ mode
-    /// still needs a guard for sequential processing.
-    fn acquire_pq_only(collection: &'a Collection, mode: StorageMode) -> Self {
-        Self {
-            sq8: None,
-            binary: None,
-            pq: matches!(mode, StorageMode::ProductQuantization)
-                .then(|| collection.pq_cache.write()),
-        }
-    }
-}
 
 impl Collection {
     /// Inserts or updates points in the collection.
