+++ crates/velesdb-core/src/collection/core/mod.rs
@@ -6,9 +6,9 @@
 //! - Index management: `create_property_index`, `create_range_index`,
 //!   `list_indexes`, `drop_index`
 
+mod bulk_import;
 mod crud;
 mod crud_bulk;
-mod bulk_import;
 mod crud_helpers;
 mod crud_histogram;
 mod crud_indexing;
